Add feature to select which subset of config is cloned/merged/deployed
Adds --include and --exclude CLI parameters to specify which categories
of config data should be part of the operation. These parameters are
mutually exclusive and take CSV lists of category names, which are
documented in the help strings. When using --include only those
categories are used. With --exclude it's all but what's specified.

Update unit tests for new functionality.
Update README.md with new help strings.

Fixes #5