add migrate command
Add a new command "migrate" which will create a new backup from an existing one, optionally stripping out items with corp dependencies, or pulling the corp dependencies into the new backup.

migrate takes the following params:

dest-corp: corp to migrate to
file: backup file to migrate
out: new file to create
strip: strip any item with a dependency from the new backup, this can be done without the ability to connect to the corp the backups were originally created on
migrate-users: will preserve the site_members in the migrated backup

This also adds a new 'corp_items' section to the backup json. Deploy has been updated to handle deploying the items in the proper order, and should ignore a backup file without corp_items so this shouldn't break anything.